Create modular-multiplicative-inverse-js
Using Javascript to mimic the process for finding the modular multiplicative inverse (MMI) of a number with a given modulus using the Euclidean Algorithm iteratively. MMI is used in cryptographic implementations.

Once this is completely built up, with a coprime checker, and a logged process, I'll refactor the solution using Rust.