Create helicarrier-modular-multiplicative-inverse.rs
Modular multiplicative inverse is an important part of any cryptographic implementation. This is my first attempt at writing code in Rust that calculates the Modular multiplicative inverse of a number using the Euclidean Algorithm. I heavily commented on my code so it would provide a lot of explainers to someone who has experience programming but is new to Rust and the Euclidean Algorithm.